+++ src/Components/PageComponent.ts
@@ -1,0 +1,6 @@
+import { Component } from '../modula';
+
+export default abstract class PageComponent extends Component
+{
+    public isPageComponent: boolean = true;
+}