+++ src/Components/TemplateComponent.ts
@@ -1,0 +1,6 @@
+import { Component } from '../modula';
+
+export default abstract class TemplateComponent extends Component
+{
+    public isTemplateComponent: boolean = true;
+}